requirements.txt
mkdocs~=1.6
mkdocs-material~=9.5.24
mkdocs-bibtex~=2.15.0
mkdocs-link-embeds-plugin
mkdocs-git-revision-date-localized-plugin
mkdocs-git-committers-plugin-2
markdown-customblocks
pybtex~=0.24.0
pypandoc_binary~=1.13
